requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
curl-cffi==0.11.3
datasets==3.6.0
db==0.1.1
db-sqlite3==0.0.1
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
duckduckgo-search==6.2.0
exceptiongroup==1.3.0
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.6
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.6
openai==1.86.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyparsing==3.2.3
pyreqwest-impersonate==0.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.62
zstandard==0.23.0
flask==2.3.2
flask-cors==3.0.10
Flask
requests
google-search-results
yfinance
beautifulsoup4
google-generativeai
yfinance
duckduckgo-search
langchain-google-genai
pydantic
